refactor: 層を分ける意味がないので削除した
