+++ lib/actions/activity_manager.go
@@ -3,7 +3,9 @@ package actions
 import (
 	"fmt"
 
+	gc "github.com/kijimaD/ruins/lib/components"
 	"github.com/kijimaD/ruins/lib/logger"
+	"github.com/kijimaD/ruins/lib/turns"
 	w "github.com/kijimaD/ruins/lib/world"
 	ecs "github.com/x-hgg-x/goecs/v2"
 )
@@ -22,6 +24,65 @@ func NewActivityManager(logger *logger.Logger) *ActivityManager {
 		currentActivities: make(map[ecs.Entity]*Activity),
 		logger:            logger,
 	}
+}
+
+// Execute は指定されたアクション（アクティビティ）を実行する
+// 即座実行アクション（移動、攻撃等）も継続アクション（休息等）も統一的に処理
+func (am *ActivityManager) Execute(activityType ActivityType, params ActionParams, world w.World) (*ActionResult, error) {
+	am.logger.Debug("アクション実行開始",
+		"type", activityType.String(),
+		"actor", params.Actor)
+
+	// アクティビティを作成
+	activity, err := am.createActivity(activityType, params, world)
+	if err != nil {
+		return &ActionResult{
+			Success:      false,
+			ActivityType: activityType,
+			Message:      err.Error(),
+		}, err
+	}
+
+	// アクティビティを開始
+	if err := am.StartActivity(activity, world); err != nil {
+		return &ActionResult{
+			Success:      false,
+			ActivityType: activityType,
+			Message:      err.Error(),
+		}, err
+	}
+
+	// 即座実行アクション（1ターン）の場合は即座に処理
+	if activity.TurnsTotal == 1 {
+		// ターン処理実行
+		am.ProcessTurn(world)
+
+		// ターン管理システムに移動コストを通知
+		am.consumeMoveCost(world, activityType, params.Actor)
+
+		// 結果を確認
+		currentActivity := am.GetCurrentActivity(params.Actor)
+		if currentActivity == nil || currentActivity.IsCompleted() {
+			return &ActionResult{
+				Success:      true,
+				ActivityType: activityType,
+				Message:      "アクション完了",
+			}, nil
+		} else if currentActivity.IsCanceled() {
+			return &ActionResult{
+				Success:      false,
+				ActivityType: activityType,
+				Message:      currentActivity.CancelReason,
+			}, fmt.Errorf("アクション失敗: %s", currentActivity.CancelReason)
+		}
+	}
+
+	// 継続アクションの場合は開始成功を返す
+	return &ActionResult{
+		Success:      true,
+		ActivityType: activityType,
+		Message:      "アクション開始",
+	}, nil
 }
 
 // StartActivity は新しいアクティビティを開始する
@@ -253,4 +314,164 @@ func (am *ActivityManager) validateResume(activity *Activity, world w.World) err
 
 	// 基本要件を再チェック
 	return am.validateBasicRequirements(activity)
+}
+
+// createActivity はActivityTypeとパラメータからアクティビティを作成する
+func (am *ActivityManager) createActivity(activityType ActivityType, params ActionParams, world w.World) (*Activity, error) {
+	switch activityType {
+	case ActivityMove:
+		if params.Destination == nil {
+			return nil, fmt.Errorf("移動先が指定されていません")
+		}
+		characterAP := am.getEntityMaxAP(params.Actor, world)
+		duration := CalculateRequiredTurns(ActivityMove, characterAP)
+		activity := NewActivity(ActivityMove, params.Actor, duration)
+		activity.Position = params.Destination
+		return activity, nil
+
+	case ActivityAttack:
+		if params.Target == nil {
+			return nil, fmt.Errorf("攻撃対象が指定されていません")
+		}
+		characterAP := am.getEntityMaxAP(params.Actor, world)
+		duration := CalculateRequiredTurns(ActivityAttack, characterAP)
+		activity := NewActivity(ActivityAttack, params.Actor, duration)
+		activity.Target = params.Target
+		return activity, nil
+
+	case ActivityPickup:
+		return &Activity{
+			Actor:      params.Actor,
+			Type:       ActivityPickup,
+			State:      ActivityStateRunning,
+			TurnsLeft:  1,
+			TurnsTotal: 1,
+			Message:    "アイテムを拾得中...",
+			Logger:     logger.New(logger.CategoryAction),
+		}, nil
+
+	case ActivityWarp:
+		return &Activity{
+			Actor:      params.Actor,
+			Type:       ActivityWarp,
+			State:      ActivityStateRunning,
+			TurnsLeft:  1,
+			TurnsTotal: 1,
+			Message:    "ワープ中...",
+			Logger:     logger.New(logger.CategoryAction),
+		}, nil
+
+	case ActivityRest:
+		duration := params.Duration
+		if duration <= 0 {
+			characterAP := am.getEntityMaxAP(params.Actor, world)
+			duration = CalculateRequiredTurns(ActivityRest, characterAP)
+		}
+		return &Activity{
+			Actor:      params.Actor,
+			Type:       ActivityRest,
+			State:      ActivityStateRunning,
+			TurnsLeft:  duration,
+			TurnsTotal: duration,
+			Message:    "休息中...",
+			Logger:     logger.New(logger.CategoryAction),
+		}, nil
+
+	case ActivityWait:
+		duration := params.Duration
+		if duration <= 0 {
+			characterAP := am.getEntityMaxAP(params.Actor, world)
+			duration = CalculateRequiredTurns(ActivityWait, characterAP)
+		}
+		return &Activity{
+			Actor:      params.Actor,
+			Type:       ActivityWait,
+			State:      ActivityStateRunning,
+			TurnsLeft:  duration,
+			TurnsTotal: duration,
+			Message:    "待機中...",
+			Logger:     logger.New(logger.CategoryAction),
+		}, nil
+
+	case ActivityRead:
+		duration := params.Duration
+		if duration <= 0 {
+			characterAP := am.getEntityMaxAP(params.Actor, world)
+			duration = CalculateRequiredTurns(ActivityRead, characterAP)
+		}
+		activity := NewActivity(ActivityRead, params.Actor, duration)
+		activity.Target = params.Target
+		return activity, nil
+
+	case ActivityCraft:
+		duration := params.Duration
+		if duration <= 0 {
+			characterAP := am.getEntityMaxAP(params.Actor, world)
+			duration = CalculateRequiredTurns(ActivityCraft, characterAP)
+		}
+		activity := NewActivity(ActivityCraft, params.Actor, duration)
+		activity.Target = params.Target
+		return activity, nil
+
+	default:
+		return nil, fmt.Errorf("%w: %v", ErrUnsupportedActivity, activityType)
+	}
+}
+
+// consumeMoveCost はターン管理システムに移動コストを通知する
+func (am *ActivityManager) consumeMoveCost(world w.World, activityType ActivityType, actor ecs.Entity) {
+	if world.Resources.TurnManager == nil {
+		am.logger.Warn("TurnManagerリソースが見つかりません")
+		return
+	}
+
+	turnManager := world.Resources.TurnManager.(*turns.TurnManager)
+	cost, actionName := GetActivityCost(activityType)
+
+	if actor.HasComponent(world.Components.Player) {
+		turnManager.ConsumePlayerMoves(actionName, cost)
+	} else {
+		success := turnManager.ConsumeActionPoints(world, actor, actionName, cost)
+		if !success {
+			am.logger.Debug("AI移動コスト消費失敗", "actor", actor, "cost", cost)
+		}
+	}
+
+	am.logger.Debug("移動コスト消費",
+		"activity", activityType.String(),
+		"cost", cost,
+		"actor", actor,
+		"isPlayer", actor.HasComponent(world.Components.Player))
+}
+
+// getEntityMaxAP はエンティティの最大AP値を取得する
+func (am *ActivityManager) getEntityMaxAP(entity ecs.Entity, world w.World) int {
+	if turnBasedComponent := world.Components.TurnBased.Get(entity); turnBasedComponent != nil {
+		turnBased := turnBasedComponent.(*gc.TurnBased)
+		return turnBased.AP.Max
+	}
+	am.logger.Debug("TurnBasedコンポーネントが見つからない", "entity", entity)
+	return 100 // デフォルトAP値
+}
+
+// ActionParams はアクション実行時のパラメータを表す
+type ActionParams struct {
+	Actor       ecs.Entity   // アクションを実行するエンティティ
+	Target      *ecs.Entity  // 対象エンティティ（攻撃等で使用）
+	Destination *gc.Position // 対象位置（移動等で使用）
+	Duration    int          // 継続時間（休息、待機等で使用）
+	Reason      string       // 理由（待機等で使用）
+}
+
+// ActionResult はアクション実行結果を表す
+type ActionResult struct {
+	Success      bool         // 実行成功/失敗
+	ActivityType ActivityType // 実行されたアクティビティタイプ
+	Message      string       // 結果メッセージ
+}
+
+// GetActivityCost はアクティビティタイプに応じたコストと名前を返す
+func GetActivityCost(activityType ActivityType) (int, string) {
+	info := GetActivityInfo(activityType)
+	return info.ActionPointCost, info.Name
 }
